Fix crashing on close, remove file that somehow magically appeared?

+++ Entities/Player.Rendering.cs
@@ -12,6 +12,7 @@ namespace ClassicalSharp {
 		protected internal int PlayerTextureId = -1, MobTextureId = -1;
 		
 		public override void Despawn() {
+			if( api == null ) return;
 			api.DeleteTexture( ref PlayerTextureId );
 			api.DeleteTexture( ref nameTex.ID );
 		}
